Support for exporting the reserved FITS DATE keyword (plus a bunch of bug fixes to make it work)

+++ base/drms/libs/api/drms_types.h
@@ -147,9 +147,11 @@ typedef struct DRMS_Value
 #define DRMS_MISSING_STRING   ("")
 /** \brief DRMS time missing value */
 #define DRMS_MISSING_TIME     (-211087684832.184) 
-/* equal to '-4712.01.01_12:00:00.000_UT' which is the time value used
+/* equal to '-4712.01.01_12:00:00.000_TDT' which is the time value used
    missing in the MDI/SDS system. */
-#define _DRMS_IS_T_MISSING(v) (isnan(v) || (v < DRMS_MISSING_TIME + 10.0e-5 && v > DRMS_MISSING_TIME - 10.0e-5))
+/* Truncate the fractional seconds during comparison because .jsds that have time keywords whose 
+ * format fields are '0' will yield time strings that are integral */
+#define _DRMS_IS_T_MISSING(v) (isnan(v) || (v < (long long)DRMS_MISSING_TIME + 10.0e-5 && v > (long long)DRMS_MISSING_TIME - 10.0e-5))
 
 #define TSEQ_EPOCH_S MDI_EPOCH_S
 #define TSEQ_EPOCH_F MDI_EPOCH_F
